not immediately crashing with AI
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@1810 8aca7d54-2c30-db11-9de9-000461428c89

+++ tpl/ptrhashtable_tpl.h
@@ -16,7 +16,7 @@
 template<class key_t>
 class ptrhash_tpl {
 public:
-	static long hash(const key_t key)
+	static uint32 hash(const key_t key)
 	{
 		return labs((long)key);
 	}
